split shownote.py
