+++ fikanote/urls.py
@@ -5,7 +5,8 @@ admin.autodiscover()
 
 import app.views
 import app.feed
-
+import app.shownote
+ 
 urlpatterns = patterns('',
     # Examples:
     # url(r'^$', 'fikanote.views.home', name='home'),
@@ -15,7 +16,7 @@ urlpatterns = patterns('',
     url(r'^(?P<number>\d+)', app.views.episode, name='episode'),
     url(r'^agenda', app.views.agenda, name='agenda'),
     url(r'^add', app.views.add, name='add'),
-    url(r'^shownote', app.views.shownote, name='shownote'),
+    url(r'^shownote', app.shownote.shownote, name='shownote'),
     url(r'^feed/', app.feed.feed, name='feed'),
     url(r'^admin/', include(admin.site.urls)),
